+++ common/int64.hh
@@ -80,7 +80,7 @@ namespace csl
         inline const unsigned char * ucharp_data() const { return reinterpret_cast<const unsigned char *>(&value_); }
 
         /** @brief returns the size of the variable data */
-        inline size_t var_size() const { return sizeof(value_); }
+        inline uint64_t var_size() const { return sizeof(value_); }
 
         /* conversions to other types */
         /**
@@ -166,7 +166,7 @@ namespace csl
         this function assumes that v has enough space. sz will be updated to indicate, how
         many bytes were stored.
          */
-        bool to_binary(unsigned char * v, size_t & sz) const;
+        bool to_binary(unsigned char * v, uint64_t & sz) const;
 
         /**
         @brief convert to (ptr+size)
@@ -177,7 +177,7 @@ namespace csl
         this function assumes that v has enough space. sz will be updated to indicate, how
         many bytes were stored.
          */
-        bool to_binary(void * v, size_t & sz) const;
+        bool to_binary(void * v, uint64_t & sz) const;
 
         /**
         @brief stores the content of this instance to an XDR stream
@@ -295,7 +295,7 @@ namespace csl
 
         this function delegates the conversion task to the common::binry class
          */
-        bool from_binary(const unsigned char * v,size_t sz);
+        bool from_binary(const unsigned char * v,uint64_t sz);
 
         /**
         @brief convert a memory region
@@ -304,7 +304,7 @@ namespace csl
 
         this function delegates the conversion task to the common::binry class
          */
-        bool from_binary(const void * v,size_t sz);
+        bool from_binary(const void * v,uint64_t sz);
 
         /**
         @brief read the content of this instance from an XDR stream
